refactor(pr08): move fieldBuilders to @fields/ (G5g.4)
git mv src/helpers/fieldBuilders.mts -> src/fields/fieldBuilders.mts. Updates the 9 importers (Item/Weapon/Physical/Equippable/Cursable system models + Active Effect / Material / Secret system models + CoreMixin DocumentSystemModel) from '@helpers/fieldBuilders.mjs' to '@fields/fieldBuilders.mjs'.

Also drops the dead field-builder re-exports from src/helpers/index.mts (only LogHelper was ever consumed via that barrel; the field-builder re-exports had zero importers).
