+++ src/entities/activeEffects/secret/data/SecretSystemModel.mts
@@ -1,5 +1,5 @@
 import { ActiveEffectSystemModel } from '@effects/BaseActiveEffect/data/ActiveEffectSystemModel.mjs';
-import { requiredBooleanField } from '@helpers/fieldBuilders.mjs';
+import { requiredBooleanField } from '@fields/fieldBuilders.mjs';
 
 import type { SecretSystemData } from './SecretSystemData.mjs';
