+++ src/entities/items/components/Physical/data/PhysicalItemSystemModel.mts
@@ -1,13 +1,13 @@
 import { SIZES } from '@constants/sizes.mjs';
 import { IdentifiableSchemaMixin } from '@ec/Identifiable/data/index.mjs';
-import { SectionField } from '@fields/SectionField.mjs';
 import {
   optionalNumberField,
   optionalStringField,
   requiredBooleanField,
   requiredNumberField,
   useDnd35eField,
-} from '@helpers/fieldBuilders.mjs';
+} from '@fields/fieldBuilders.mjs';
+import { SectionField } from '@fields/SectionField.mjs';
 import { ItemSystemModel } from '@items/baseItem/data/index.mjs';
 import { PriceField } from '@settings/currency/PriceField.mjs';
